+++ packages/examples/test/concurrentCounter.ts
@@ -112,9 +112,6 @@ class ConcurrentCounter extends OffchainStateContract {
   public applyAction(action: Action) {
     const { id, by } = action.payload;
 
-    Circuit.log(this.lastUpdatedOffchainState);
-
-    Circuit.log('getting counter');
     const [counter] = this.getCounter(id);
 
     const newCounter = Circuit.if(
@@ -123,7 +120,6 @@ class ConcurrentCounter extends OffchainStateContract {
       safeUint64Sub(counter, by)
     );
 
-    Circuit.log('setting counter')
     this.setCounter(id, newCounter);
   }
 
@@ -143,6 +139,8 @@ class ConcurrentCounter extends OffchainStateContract {
     });
 
     const currentRootHash = this.root.getRootHash();
+
+    this.resetLastUpdatedOffchainState();
 
     const { actionsHash: newActionsHash, state: newRootHash } =
       this.withRollingState(() =>
