+++ packages/examples/test/doubleCounter.test.ts
@@ -6,13 +6,13 @@
 import { Key } from '@zkfs/contract-api';
 import { AccountUpdate, MerkleMap, Poseidon, UInt64 } from 'snarkyjs';
 
-import Counter2 from './counter2.js';
+import DoubleCounter from './doubleCounter.js';
 import describeContract, { withTimer } from './describeContract.js';
 
 Error.stackTraceLimit = 10_000_000;
 
 // eslint-disable-next-line jest/require-hook
-describeContract<Counter2>('counter', Counter2, (context) => {
+describeContract<DoubleCounter>('doubleCounter', DoubleCounter, (context) => {
   async function localDeploy() {
     const {
       deployerAccount,
@@ -21,8 +21,6 @@ describeContract<Counter2>('counter', Counter2, (context) => {
       zkApp,
       contractApi,
     } = context();
-
-    zkApp.lastUpdatedOffchainState = undefined;
 
     const tx = await withTimer(
       'transaction',
@@ -46,9 +44,6 @@ describeContract<Counter2>('counter', Counter2, (context) => {
 
   it('correctly updates the count state on the `Counter` smart contract', async () => {
     expect.assertions(1);
-
-    // eslint-disable-next-line @typescript-eslint/no-use-before-define
-    const { map, count1, count2 } = manualMapTesting();
 
     const { senderAccount, senderKey, zkApp, contractApi } = context();
 
@@ -89,7 +84,6 @@ describeContract<Counter2>('counter', Counter2, (context) => {
       tx: tx1.toPretty(),
     });
 
-
     console.log(
       'Counter.update() the second time, updating the offchain state...'
     );
@@ -120,42 +114,24 @@ describeContract<Counter2>('counter', Counter2, (context) => {
       // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
       tx: tx2.toPretty(),
     });
-    map.set(count1, Poseidon.hash(UInt64.from(2).toFields()));
-    map.set(count2, Poseidon.hash(UInt64.from(4).toFields()));
-    expect(zkApp.offchainStateRootHash.get().toString()).toBe(map.getRoot().toString())
+
+    // eslint-disable-next-line @typescript-eslint/no-use-before-define
+    const { map } = manualMapTesting();
+
+    expect(zkApp.offchainStateRootHash.get().toString()).toBe(
+      map.getRoot().toString()
+    );
   });
 });
 
 function manualMapTesting() {
-  //console.log(UInt64.from(0));
-
   const map = new MerkleMap();
-  //console.log('root before init should be', map.getRoot().toString());
 
   const count1 = Key.fromString('count1').toField();
-  map.set(count1, Poseidon.hash(UInt64.from(0).toFields()));
+  map.set(count1, Poseidon.hash(UInt64.from(2).toFields()));
 
   const count2 = Key.fromString('count2').toField();
-  map.set(count2, Poseidon.hash(UInt64.from(0).toFields()));
-  //console.log('root after init should be', map.getRoot().toString());
+  map.set(count2, Poseidon.hash(UInt64.from(4).toFields()));
 
-  map.set(count1, Poseidon.hash(UInt64.from(1).toFields()));
-  // console.log(
-  //   'root after setting count1 should be',
-  //   map.getRoot().toString()
-  // );
-
-  map.set(count2, Poseidon.hash(UInt64.from(2).toFields()));
-  // console.log(
-  //   'root after setting count2 should be',
-  //   map.getRoot().toString()
-  // );
-  // console.log('final root after update should be', map.getRoot().toString());
-
-  const map2 = new MerkleMap();
-  map2.set(count1, Poseidon.hash(UInt64.from(0).toFields()));
-  map2.set(count2, Poseidon.hash(UInt64.from(2).toFields()));
-  //console.log('root if only count2 was set', map2.getRoot().toString());
   return { map, count1, count2 };
 }
-
